Mobile css and tablet css
Created a separate style sheet for mobile and tablet.
For mobile css and tablet css- only changes I made is removing the banner, and the layout and the background to indicate which mode.

ISSUE: I am not sure if the website is visible on the phone or laptop. In the chrome developer tool, it says no working, but resizing my monitor does work.